Moved CLI into its own module with submodules
Less chaos in one file
